Moved CLusterClient and ClusterServer classes to own files
