+++ packages/matter.js/src/cluster/server/GroupsServer.ts
@@ -13,8 +13,8 @@ import { Fabric } from "../../fabric/Fabric.js";
 import { SessionType } from "../../codec/MessageCodec.js";
 import { ScenesManager } from "./ScenesServer.js";
 import { IdentifyCluster } from "../definitions/IdentifyCluster.js";
-import { ClusterServer } from "../../protocol/interaction/InteractionServer.js";
 import { NotImplementedError } from "../../common/MatterError.js";
+import { ClusterServer } from "./ClusterServer.js";
 
 /*
 TODO: If the Scenes server cluster is implemented on the same endpoint, the following extension field SHALL
