+++ packages/matter.js/src/cluster/server/OnOffServer.ts
@@ -6,7 +6,8 @@
 
 import { OnOffCluster } from "../definitions/OnOffCluster.js";
 import { AttributeInitialValues, ClusterServerHandlers } from "./ClusterServerTypes.js";
-import { ClusterServer } from "../../protocol/interaction/InteractionServer.js";
+
+import { ClusterServer } from "./ClusterServer.js";
 
 /*
 TODO: Global Cluster fields needs to be added also here because, as discussed, based on the implementation.
