+++ packages/matter.js/test/cluster/ClusterFactoryTest.ts
@@ -10,8 +10,8 @@ import { TlvUInt8 } from "../../src/tlv/TlvNumber.js";
 import { TlvNoArguments } from "../../src/tlv/TlvNoArguments.js";
 import { BitFlag, BitFlags, TypeFromPartialBitSchema } from "../../src/schema/BitmapSchema.js";
 import { TlvField, TlvObject } from "../../src/tlv/TlvObject.js";
-import { ClusterServer } from "../../src/protocol/interaction/InteractionServer.js";
 import { ClusterServerHandlers } from "../../src/cluster/server/ClusterServerTypes.js";
+import { ClusterServer } from "../../src/cluster/server/ClusterServer.js";
 
 enum Feature {
     Extended = "Extended",
